requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.6
cryptography==36.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-allauth==0.46.0
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-extensions==3.1.5
django-model-utils==4.2.0
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==9.9.0
idna==3.3
iniconfig==1.1.1
jsonfield==3.1.0
kombu==5.2.2
mock==4.0.3
mysqlclient==2.1.0
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
prompt-toolkit==3.0.23
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
